Create Evaluate Division
Basically the idea is to use BFS or DFS, how?
say we are a/b and b/c and we need to find a/c, so lets make a weighted graph where a and b are nodes and its value is the weight of the edge connecting a and b, now one thing to note is that, to get the value of  a/c there should definitely exist a path from a to c, so if we have a path from a to b and then from b to c and if we multiply the weights of all the edges present in the path  then, it would be something like (a/b) * (b/ c) , here b gets cancelled, and we ultimately get a/c, also as the graph will be bi directional, what we want is a/b if it is 2.0 then b/a should be 0.5, i.e the reciprocal of a/b, so when we build the graph the edge weight will be different for a->b and b->a. Atlast if there is no path from src to dest then we print -1.